# Remove double decoding in unmarshalSingleValue()